Added a log file
There's no good way of allowing the user to "See" the output of the
program AND be able to input (that I know of. Obviously, if there are
better ways of doing this, let me know).
The solution that I've come up with will write any statuses of the
elevators to a log and allow the user to shoot commands and stuff. I
mean, right now it isn't working, but that's the plan.