Got the elevator to operate in the way I want!
However, there are massive changes that still need to be made. Now that
the framework is more or less in place, I need to do considerable
planning around the rules of how elevators' commands are called and
queued. For the time being, the elevator sort of works, so that's good.